# src/main/java/com/bmps/difftool/exception/ServerException.java
package com.bmps.difftool.exception;

import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.ResponseStatus;

@ResponseStatus(code = HttpStatus.INTERNAL_SERVER_ERROR)
public class ServerException extends RuntimeException {
    private Object error = null;

    public ServerException error(Object error) {
        this.error = error;
        return this;
    }

    public Object getError() {
        return error;
    }

    public void setError(Object error) {
        this.error = error;
    }
}